+++ packages/bridge-controller/src/bridge-controller.test.ts
@@ -2578,6 +2578,7 @@ describe('BridgeController', function () {
         // Use the actual Solana mock which already has string trade type
         const btcQuoteResponse = mockBridgeQuotesSolErc20V1.map((quote) => ({
           ...quote,
+          quoteId: quote.quote.requestId,
           quote: {
             ...quote.quote,
             srcChainId: ChainId.BTC,
@@ -2671,11 +2672,11 @@ describe('BridgeController', function () {
         expect(quotes[1].nonEvmFeesInNative).toBeUndefined();
         expect(consoleErrorSpy).toHaveBeenCalledTimes(2);
         expect(consoleErrorSpy).toHaveBeenCalledWith(
-          'Failed to compute non-EVM fees for quote in bip122:000000000019d6689c085ae165831e93:',
+          'Failed to compute non-EVM fees for quote 5cb5a527-d4e4-4b5e-b753-136afc3986d3',
           new Error('Failed to compute fees'),
         );
         expect(consoleErrorSpy).toHaveBeenCalledWith(
-          'Failed to compute non-EVM fees for quote in bip122:000000000019d6689c085ae165831e93:',
+          'Failed to compute non-EVM fees for quote 12c94d29-4b5c-4aee-92de-76eee4172d3d',
           new Error('Failed to compute fees'),
         );
       },